Rename runtime event action route module
